+++ CHANGELOG.md
@@ -12,5 +12,6 @@ This changelog is initialized in release 1.0.0
 ### Added
 * CHANGELOG.md
 * Columns argument to ModelRepositoryInterface::all to be compatible with the underlying Eloquent model
+* FQN options for contract and repository to ModelRepositoryMakeCommand
 
 [Unreleased]: https://github.com/wimski/laravel-model-repositories/compare/v1.0.0...master